refactor(ext2): `block_group_descriptor.rs` -> `bgd.rs`
